Berhasil Sistem Penjualan di Admin, tapi masih ingin modif struk

+++ app/Livewire/Pos/ManualItemForm.php
@@ -30,7 +30,11 @@ class ManualItemForm extends Component
     {
         $this->validate();
 
-        $numericPrice = (float) $this->price;
+        // Konversi nilai mentah dari AutoNumeric menjadi integer yang benar.
+        // Input "125.000" dari AutoNumeric akan menjadi string "125000".
+        // Kita pastikan ini adalah integer.
+        $numericPrice = (int) $this->price;
+
         if ($numericPrice <= 0) {
             $this->addError('price', 'Harga harus lebih besar dari 0.');
             return;
@@ -40,18 +44,18 @@ class ManualItemForm extends Component
             'id'     => 'manual-' . uniqid(),
             'name'   => $this->name,
             'qty'    => 1,
-            'price'  => $numericPrice,
+            'price'  => $numericPrice, // Gunakan harga yang sudah pasti integer
             'weight' => 1,
             'options' => [
                 'source_type'           => 'manual',
                 'product_discount'      => 0,
                 'product_discount_type' => 'fixed',
-                'sub_total'             => $numericPrice,
+                'sub_total'             => $numericPrice, // Gunakan harga yang sudah pasti integer
                 'code'                  => '-',
                 'stock'                 => 1,
                 'unit'                  => 'unit',
                 'product_tax'           => 0,
-                'unit_price'            => $numericPrice,
+                'unit_price'            => $numericPrice, // Gunakan harga yang sudah pasti integer
                 'hpp'                   => 0,
                 'discount'              => 0,
                 'tax'                   => 0,
@@ -59,6 +63,7 @@ class ManualItemForm extends Component
         ]);
 
         $this->reset(['name', 'price']);
+        // Ganti event 'cartUpdated' agar lebih spesifik targetnya ke komponen Checkout
         $this->dispatch('cartUpdated')->to(\App\Livewire\Pos\Checkout::class);
         $this->dispatch('showSuccess', ['message' => 'Item manual ditambahkan!']);
     }
